fixup! hw/arm: Add support for Hercules SoC
