+++ hw/arm/Makefile.objs
@@ -54,4 +54,4 @@ obj-$(CONFIG_FSL_IMX7) += fsl-imx7.o mcimx7d-sabre.o
 obj-$(CONFIG_ARM_SMMUV3) += smmu-common.o smmuv3.o
 obj-$(CONFIG_FSL_IMX6UL) += fsl-imx6ul.o mcimx6ul-evk.o
 obj-$(CONFIG_NRF51_SOC) += nrf51_soc.o
-obj-$(CONFIG_HERCULES) += hercules.o tms570lcxxxx.o
+obj-$(CONFIG_HERCULES) += hercules.o
